pyproject.toml
[project]
name = "ivy-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "apscheduler>=3.11.2",
    "chromadb>=1.5.5",
    "document>=1.0",
    "docx>=0.2.4",
    "langchain>=1.2.12",
    "langchain-anthropic>=1.3.5",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-google-community>=3.0.5",
    "langchain-google-genai>=4.2.1",
    "langchain-text-splitters>=1.1.1",
    "numpy>=2.4.3",
    "openai>=2.28.0",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "reportlab>=4.4.10",
    "requests>=2.32.5",
    "sentence-transformers>=5.3.0",
    "tools>=1.0.23",
    "uv==0.10.10",
    "uvicorn[standard]>=0.41.0",
]
